requirements.txt
Django==2.0.5
django-tastypie==0.14.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
